requirements.txt
Django==3.2
gunicorn==20.1.0
whitenoise==5.2.0

psycopg2-binary==2.8.6

django-redis==4.12.1
celery==5.0.5

flake8==3.9.0
spotipy
python-dotenv
django-debug-toolbar

djangorestframework
markdown
django-filter